+++ Assets/Scripts/Text_inout.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 07736478c4029914a839629b4864a0b1
+guid: 03c5a02b5aa56ab47955fc21e5f51cea
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
